+++ src/exceptions/exception.ts
@@ -3,6 +3,7 @@ export type ExceptionCategory = (typeof EXCEPTION_CATEGORIES)[number];
 
 export interface Exception {
   readonly category: ExceptionCategory;
+  readonly defaultEnabled?: true;
   enabled?: boolean;
   readonly name: string;
   readonly userDefined?: true;